feat: 強化資料安全、復原機制與持續整合

+++ YTLiveChatCatcher/Common/Utils/ChatStatsCalculator.cs
@@ -135,9 +135,9 @@ public static partial class ChatStatsCalculator
     /// </summary>
     /// <param name="purchaseAmountText">字串，購買金額文字</param>
     /// <param name="currencySymbol">out 字串，貨幣符號（例如 "NT$"）</param>
-    /// <param name="amount">out double，數字金額</param>
+    /// <param name="amount">out decimal，數字金額</param>
     /// <returns>布林值，是否成功解析出金額</returns>
-    public static bool TryParsePurchaseAmount(string purchaseAmountText, out string currencySymbol, out double amount)
+    public static bool TryParsePurchaseAmount(string purchaseAmountText, out string currencySymbol, out decimal amount)
     {
         Match match = PurchaseAmountRegex().Match(purchaseAmountText);
 
@@ -156,11 +156,11 @@ public static partial class ChatStatsCalculator
         currencySymbol = symbol == "$" ? "NT$" : symbol;
 
         // 2026/9 修正：正規表示式擷取出的金額字串固定是英式格式（逗號千分位、句點小數點，
-        // 見上方文件註解），但 double.TryParse 沒指定 NumberStyles／CultureInfo 時會用呼叫端執行緒
+        // 見上方文件註解），但 decimal.TryParse 沒指定 NumberStyles／CultureInfo 時會用呼叫端執行緒
         // 的 CurrentCulture 判讀，受 Windows 地區設定影響——若使用者的地區設定把逗號當小數點、
         // 句點當千分位（常見於多數歐洲地區設定），千元以上的金額會解析失敗，收益統計因此失準。
         // 固定用 InvariantCulture 判讀，不受執行環境的地區設定影響。
-        return double.TryParse(
+        return decimal.TryParse(
             match.Groups["amount"].Value,
             NumberStyles.AllowThousands | NumberStyles.AllowDecimalPoint,
             CultureInfo.InvariantCulture,
